delete empty file ssh.go, and rename ssh_v2.go to ssh.go
